+++ spec/services/person/filterer_spec.rb
@@ -12,13 +12,16 @@ describe Person::Filterer do
 
   describe '.filter_classes' do
     it 'returns an empty array' do
-      expect(described_class.filter_classes).to eq [Person::Filter::EmailAddressValid, Person::Filter::PhoneNumberValid, Person::Filter::UpdatedAt]
+      expect(described_class.filter_classes).to eq [Person::Filter::EmailAddressValid,
+                                                    Person::Filter::PhoneNumberValid,
+                                                    Person::Filter::UpdatedAt,
+                                                    Person::Filter::WildcardSearch]
     end
   end
 
   describe '.filter_params' do
     it 'returns an empty array' do
-      expect(described_class.filter_params).to eq [:email_address_valid, :phone_number_valid, :updated_at]
+      expect(described_class.filter_params).to eq [:email_address_valid, :phone_number_valid, :updated_at, :wildcard_search]
     end
   end